synthetic/requirements.txt
# Rendering Pipeline Requirements
# Phase 2: Synthetic Data Generation

# Core dependencies
numpy>=1.24.0,<2.0.0
Pillow>=10.0.0
requests>=2.31.0  # For OSM data fetching

# Vector tile handling
pmtiles>=3.3.0
mapbox-vector-tile>=2.0.1

# Progress bars
tqdm>=4.65.0

# Optional: MapLibre native rendering (if available)
# Note: maplibre-native is not available as a standard pip package
# If you want to use it, you'll need to build from source:
# https://github.com/maplibre/maplibre-native

# Optional: Additional image processing (for aging effects)
opencv-python>=4.8.0  # For advanced image manipulation
scipy>=1.11.0  # For image filters and transformations

# Development dependencies (optional)
# pytest>=7.4.0
# pytest-cov>=4.1.0
# black>=23.7.0
# flake8>=6.1.0
# mypy>=1.5.0
